+++ src/commands/commandManager.ts
@@ -1,7 +1,8 @@
+import { extractVariableName, getVscodeConfigTarget } from 'utils/variables';
 import * as vscode from 'vscode';
-import { FileWatcher } from '../watchers/fileWatcher';
 import I18n from '../translations';
 import { VariablePanelManager } from '../ui/panels/variablePanel';
+import { FileWatcher } from '../watchers/fileWatcher';
 
 export class CommandManager {
   private fileWatcher: FileWatcher;
@@ -12,8 +13,7 @@ export class CommandManager {
     this.variablePanelManager = variablePanelManager;
   }
 
-
-  public checkVariables() {
+  public async checkVariables() {
     vscode.window.showInformationMessage(I18n.__('variable.checkingVariables'));
     const editor = vscode.window.activeTextEditor;
     if (editor) {
@@ -38,4 +38,53 @@ export class CommandManager {
     }
   }
 
+  public async saveVariable(diagnosticMessage: string) {
+    const variable = extractVariableName(diagnosticMessage);
+    const activeEditor = vscode.window.activeTextEditor;
+
+    if (!variable) {
+      vscode.window.showWarningMessage(I18n.__('error.variableNotFound'));
+      return;
+    }
+
+    if (!activeEditor) {
+      vscode.window.showWarningMessage(I18n.__('error.noActiveEditor'));
+      return;
+    }
+
+    const filePath = activeEditor.document.uri.fsPath;
+    const workspaceFolder = vscode.workspace.getWorkspaceFolder(activeEditor.document.uri);
+
+    if (!workspaceFolder) {
+      vscode.window.showWarningMessage(I18n.__('warning.noWorkspaceFolder'));
+    }
+
+    const target = getVscodeConfigTarget(activeEditor);
+
+    const targetTranslation = (target === vscode.ConfigurationTarget.WorkspaceFolder) && workspaceFolder
+      ? I18n.__('quickFix.workspaceTarget')
+      : I18n.__('quickFix.globalTarget');
+
+    const config = (target === vscode.ConfigurationTarget.WorkspaceFolder) && workspaceFolder
+      ? vscode.workspace.getConfiguration('jinja2-html-enhancer', workspaceFolder.uri)
+      : vscode.workspace.getConfiguration('jinja2-html-enhancer');
+
+    const currentVariables: { [key: string]: string[] } = config.get('customVariables', {});
+
+    if (currentVariables[filePath] && currentVariables[filePath].includes(variable)) {
+      vscode.window.showWarningMessage(I18n.__('warning.variableExists', { variable, target: targetTranslation }));
+      return;
+    }
+
+    const variableArray = currentVariables[filePath] || [];
+    variableArray.push(variable);
+
+    try {
+      await config.update('customVariables', { ...currentVariables, [filePath]: variableArray }, target);
+      vscode.window.showInformationMessage(I18n.__('quickFix.save', { variable, target: targetTranslation }));
+      await this.checkVariables();
+    } catch (error) {
+      vscode.window.showErrorMessage(I18n.__('error.variableNotSaved', { variable, error: String(error) }));
+    }
+  }
 }