+++ src/diagnostics/diagnosticsManager.ts
@@ -35,12 +35,12 @@ export class DiagnosticsManager {
             I18n.__('variable.variableNotSet', { variable }),
             vscode.DiagnosticSeverity.Warning
           );
+          diagnostic.code = 'JHE0001';
           diagnostics.push(diagnostic);
         }
       }
     });
 
-    this.diagnosticCollection.set(document.uri, diagnostics);
     this.diagnosticCollection.set(document.uri, diagnostics);
     this._onDidUpdateDiagnostics.fire({ usedVariables, setVariables });
   }