+++ src/utils/variables.ts
@@ -1,0 +1,16 @@
+import * as vscode from 'vscode';
+
+export function extractVariableName(diagnosticMessage: string): string | null {
+  const match = diagnosticMessage.match(/'([^']+)'/);
+  return match ? match[1] : null;
+}
+
+export const getVscodeConfigTarget = (activeEditor: vscode.TextEditor) => {
+  const workspaceFolder = vscode.workspace.getWorkspaceFolder(activeEditor.document.uri);
+
+  const vscodeConfigTarget = workspaceFolder
+    ? vscode.ConfigurationTarget.WorkspaceFolder
+    : vscode.ConfigurationTarget.Global;
+
+  return vscodeConfigTarget;
+};