+++ src/watchers/fileWatcher.ts
@@ -1,6 +1,8 @@
+import { getVscodeConfigTarget } from 'utils/variables';
 import * as vscode from 'vscode';
 import { DiagnosticsManager } from '../diagnostics/diagnosticsManager';
 import { analyzeNestedStructures, extractVariables } from '../diagnostics/variableAnalyzer';
+import I18n from '../translations';
 
 export class FileWatcher {
   private readonly diagnosticsManager: DiagnosticsManager;
@@ -24,6 +26,24 @@ export class FileWatcher {
   }
 
   public analyzeDocument(document: vscode.TextDocument | vscode.Uri) {
+    const activeEditor = vscode.window.activeTextEditor;
+    if (!activeEditor) {
+      vscode.window.showWarningMessage(I18n.__('error.noActiveEditor'));
+      return;
+    }
+
+    const workspaceFolder = vscode.workspace.getWorkspaceFolder(activeEditor.document.uri);
+    if (!workspaceFolder) {
+      vscode.window.showWarningMessage(I18n.__('warning.noWorkspaceFolder'));
+    }
+
+    const target = getVscodeConfigTarget(activeEditor);
+    const config = (target === vscode.ConfigurationTarget.WorkspaceFolder) && workspaceFolder
+      ? vscode.workspace.getConfiguration('jinja2-html-enhancer', workspaceFolder.uri)
+      : vscode.workspace.getConfiguration('jinja2-html-enhancer');
+
+    const customVariables: { [key: string]: string[] } = config.get('customVariables', {});
+
     if (document instanceof vscode.Uri) {
       vscode.workspace.openTextDocument(document).then(this.analyzeDocument.bind(this));
       return;
@@ -33,8 +53,7 @@ export class FileWatcher {
     const { usedVariables, setVariables } = extractVariables(text);
     const nestedVariables = analyzeNestedStructures(text);
 
-    const allSetVariables = [...new Set([...setVariables, ...nestedVariables])];
-
+    const allSetVariables = [...new Set([...setVariables, ...nestedVariables, ...(customVariables ? Object.values(customVariables).flat() : [])])];
     this.diagnosticsManager.updateDiagnostics(document, usedVariables, allSetVariables);
 
     return {